requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-filter==24.2
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
kombu==5.3.7
multidict==6.0.5
packaging==24.1
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stripe==9.10.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
